+++ Tests/GeneratedTests/GeneratedTests.swift
@@ -248,6 +248,12 @@ class DiscouragedOptionalCollectionRuleGeneratedTests: SwiftLintTestCase {
     }
 }
 
+class DiscouragedVoidReturnRuleGeneratedTests: SwiftLintTestCase {
+    func testWithDefaultConfiguration() {
+        verifyRule(DiscouragedVoidReturnRule.description)
+    }
+}
+
 class DuplicateConditionsRuleGeneratedTests: SwiftLintTestCase {
     func testWithDefaultConfiguration() {
         verifyRule(DuplicateConditionsRule.description)